requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.1
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0  # <-- ДОБАВИТЬ
httpx==0.28.1
motor==3.4.0
pymongo==4.8.0
pytest==8.2.0
pytest-cov==5.0.0
pytest-asyncio==0.23.7
typing_extensions==4.15.0